// src/app/services/projects/models/project-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectDto {
  createdAt?: string;
  createdBy?: string;
  description?: string;
  endDate?: string;
  id?: string;
  key?: string;
  memberCount?: number;
  name?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  startDate?: string;
  status?: 'ACTIVE' | 'COMPLETED' | 'ARCHIVED' | 'ON_HOLD';
  taskCount?: number;
  updatedAt?: string;
}